backend/requirements.txt
langchain
langchain-groq
langgraph
fastapi
uvicorn
tavily-python
faiss-cpu
python-dotenv
pydantic
langchain-community
tiktoken
pypdf
langchain-experimental
langchain-huggingface
fastembed
sentence-transformers
pinecone-client
langchain-pinecone
pypdf
